Permettre d’accéder aux analyses réalisées par les autres membres de son Antenne

+++ app/controllers/feedbacks_controller.rb
@@ -4,7 +4,7 @@ class FeedbacksController < ApplicationController
   before_action :authenticate_expert!, if: -> { params[:access_token].present? }
 
   def create
-    @feedback = Feedback.create(feedback_params)
+    @feedback = Feedback.create(feedback_params.merge(user: current_user))
     @current_roles = current_roles
     if @feedback.persisted?
       UserMailer.delay.match_feedback(@feedback)
@@ -16,6 +16,7 @@ class FeedbacksController < ApplicationController
 
   def destroy
     feedback = retrieve_feedback
+    authorize feedback
     @feedback_id = feedback.id
     feedback.destroy!
   end
@@ -23,13 +24,11 @@ class FeedbacksController < ApplicationController
   private
 
   def feedback_params
-    params.require(:feedback).permit(:need_id, :expert_id, :user_id, :description)
+    params.require(:feedback).permit(:need_id, :description)
   end
 
   def retrieve_feedback
     safe_params = params.permit(:id)
-    feedback = Feedback.find(safe_params[:id])
-    check_current_user_access_to(feedback)
-    feedback
+    Feedback.find(safe_params[:id])
   end
 end
